refactor(frontend/i18n): move into modules
