+++ apps/frontend/app/[lang]/(auth)/login/dictionary.ts
@@ -1,5 +1,5 @@
-import { Lang } from '@/features/i18n/types'
-import { getPageDictionary } from '@/features/i18n/utils'
+import { Lang } from '@/modules/i18n/types'
+import { getPageDictionary } from '@/modules/i18n/utils'
 
 export const getDictionary = (lang: Lang) =>
 	getPageDictionary<'login'>(lang, 'login')
