+++ apps/frontend/modules/i18n/dict/en.ts
