+++ apps/frontend/modules/metadata.ts
@@ -1,5 +1,5 @@
-import { LOCALES } from '@/features/i18n/config'
-import { Lang, LangPageParams } from '@/features/i18n/types'
+import { LOCALES } from '@/modules/i18n/config'
+import { Lang, LangPageParams } from '@/modules/i18n/types'
 
 type Metadata = {
 	title: string
